requirements.txt
# Core data processing and analysis
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Web API and HTTP requests
requests>=2.31.0
python-dotenv>=1.0.0

# Machine Learning and NLP
transformers>=4.30.0
torch>=2.0.0

# Data visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Web application framework
streamlit>=1.28.0

# Terminal output formatting
colorama>=0.4.6

# Jupyter notebook support
notebook>=7.0.0

# Optional: AI/LLM integration (if using OpenAI)
openai>=1.0.0

# Note: beautifulsoup4 and wordcloud were included but not used in the current codebase
# Uncomment below if needed for future web scraping or word cloud features:
# beautifulsoup4>=4.12.0
# wordcloud>=1.9.0
